Disable rendering if url() given to fill/stroke with no fallback
The SVG spec states that if a stroke/fill is given that has a url() which
is unresolvable, and there is no fallback option given after the url(),
it's an error.  Given that prawn-svg doesn't currently handle url() colors,
we now handle this situation by treating the element as if its stroke/fill
was set to "none".

This change also implements the 'inherit' and 'currentColor' attribute values,
as well as being a bit more lenient around whitespace and capitalisation.

Resolves #59.