Remove Parser classes, implement element classes for each tag
This is a significant refactor!  Now all the parsing logic is contained
in Prawn::SVG::Elements::X, where X is the name of the tag that's being
processed.  It's a two-step process: first the element is parsed to
determine whether we want to render it or not, and then the call stack
is appended to if all is well.

This change will also let us calculate the bounding box for every draw,
which is required to do gradients.