Implement correct(er) font-family parsing
Previously we were just splitting on comma and hoping everything would go OK.
Now we have a much better parser for font-family.  It's really slow, though.
Something to optimise later if it turns out to be a problem.

The spec says "monospace" (with quotes) shouldn't match the in-built monospace
font, but this parser doesn't keep track of whether something was quoted or
not, so it's not quite perfect.  But it's better.