Implement <marker> support
This is a full implementation of <marker> except that properties specified
on the <marker> or its ancestors are not carried through into the marker
when rendering it.  This is due to a design limitation in the system
(which I'll work on removing next.)  Fortunately this feature doesn't seem
to be used by SVG generators.

Implements #49.