Separate attributes and properties into different data structures
This is step one of a multi-step migration.  Previously, attributes and
properties were all combined together in the same hash.  Now, we have
a new value object for storing properties, the concept of "computed
properties" (properties inherited from ancestor tags).  This will
change the operation of prawn-svg to be more like a proper SVG parser.