Make text state non-hierarchical
When you enter a <text> element, any co-ordinates specified in a <tspan> apply
to all text and element nodes that follow, not just ones that are children of
the tspan.  This commit creates a new state when a <text> element is entered,
and that same state is persisted until the <text> element is closed.