Massive rewrite of text processing
<text> has a very particular way of apportioning white space when you have
multiple tags inside. This is important for correct implementation of the
"rotate" attribute, which is coming next, but is also important so white
space in between nodes inside a <text> element is correct, which it hasn't
been to date.

Unfortunately the only way to do it properly has meant I've had to rewrite
the text processing so it does a parse step on all the nodes first, and then
the apply step after.  This is quite a big departure from how prawn-svg
normally works, hence the big code refactor.