Bypass processing the fake Text element class
Prawn::SVG::Elements::Text exists to wrap all the text processing
operations; it's not actually part of the element tree.  But it was
running element processing code, so any styles applied to a <text>
element would be applied twice.  That's bad if your style is relative
(like a font-size in em units.)  Fixed so it doesn't even look at
the element's attributes and delegates that to the root text component
underneath it.

Fixes #88.