Correctly handle a single requested dimension plus a viewBox
The change in #92 didn't take into account the scenario where a single
dimension was requested, and a viewBox was specified.  In this fix,
the unspecified dimension would be set to 100% of the available space
in that dimension, which doesn't feel like the right thing to do.

This change calculates the ratio of the viewBox width and height, and
uses that to calculate the missing dimension.

Hopefully a fix for #96.