Parsing fix for overly-compressed ‘a/A’ path commands
Some SVG compressors (like SVGOMG) are want to omit the spaces after the large-arc-flag and sweep-flag parameters  (so ‘A50,50 0 1 0 0,10’ becomes ‘A50 50 0 100 100’), which—understandably—confuses the default parser.
To work around this, we’ll just manually reparse ‘A’ commands as we come across them, as we know what specific format we need to expect.