Rubocop + its automated changes, no functional changes
Rubocop has been configured with a relatively light set of cops enabled.
This is more to catch out programming problems than to enforce a very
strict style guide.

The rest of the changes are made by Rubocop when run with -A.  They
should not result in any functional changes.

There are some TODOs grandfathered in by way of the .rubocop_todo.yml
file.  These can be addressed another time.