fix: make PDF pattern keys unique
Previously pattern keys were generated based on the SHA1 digest of a
number of components of the pattern.  The pattern also encodes the
position where the pattern is drawn, but the position wasn't included in
the digest, so gradients weren't showing up when reused in different
locations on the page.

This change removes the digest and puts in its place a global
incrementing counter.  Because patterns must be defined essentially
every time something with a gradient is output to the PDF, there's no
point any longer trying to make it cache.  Given that, the global
counter is an easy way to accomplish always-unique keys.