fix: correctly handle fill/stroke opacity properties
Previously, the fill-opacity and stroke-opacity properties were being
treated like the opacity property in that they were applied to all their
children.

It's correct that opacity on an element should apply to its children,
and another opacity property on a child should multiply the opacity of
its parent.  But fill-opacity and stroke-opacity are different in that
the value is inherited and any change to a child's values must
overwrite their parent's value.

This bug was spotted because `fill-opacity="0"` was being declared at
the top level of a document, with `fill-opacity="1"` on its children.
With the old code, 0 was multiplied with 1, giving an opacity of 0.
With the new code, the child's 1 value overrides the parent's 0 value.

Fixes #177