Improve syntax checking of property values
Previously, invalid values of properties would cause the property to
return to its default value in some cases, or take on the invalid value
in some other cases.  This is not what CSS is meant to do.

Now an invalid value in a property will be ignored, as if the property
was never provided.  This allows previously specified properties to
continue to be used, or for properties that inherit they will inherit
their parent's value.

This commit is part one of a larger change.  Part two will store parsed
values into the properties variable rather than the raw strings received
from the CSS.  For example, numbers will be stored as integers or
floats, and colours will be stored as a struct.