Convert property value strings to sensible values at parse time
Previously, the codebase assumed that all fields on the properties
object returned a string: the raw string that was provided in CSS or in
an element attribute.  (With one exception: recognised keywords were
stripped and normalised to lowercase.)

This meant that the property value string needed to be parsed every time
it was used throughout the codebase.

Now, property values are parsed into a value that is more easily used
throughout the codebase.  Numbers turn into floats, and there are some
new structs (Length, Percentage, Paint, FuncIRI) that hold more complex
property data.

I've also taken the opportunity to match up the parsing to be a bit
closer to what the SVG spec expects, with a few exceptions (Chrome
accepts floating point and negative RGB values, for example, so we do
too.)