Fix bug when gradient used on element without bounding box
This is a regression, which used to be handled correctly but was broken in 4da7475ee03e2df6de95c697ad8e9cbd88dfa2a7. Some elements (eg. text) don't return a valid bounding box, so they can't currently be used with `gradientUnits` set to `userSpaceOnUse`.

Because we were passing `matrix_for_bounding_box(nil)` this was causing an exception.