Support links via anchor ("a") elements
Add links using Prawn's `link_annotation`. This works with any element that defines a bounding box, and also text. We transform the bounding box into PDF page space so that the link area will tightly fit a rotated rectangle, for example.

One thing that isn't supported is adding links inside a text element, for example:

```svg
<text>Text with a <a href="...">link</a> inside.</text>
```

I believe this is valid SVG, but the current parser in prawn-svg doesn't handle this nesting, so it is beyond the scope of this change.